requirements.txt
Flask
pytesseract
Pillow
pdf2image
requests
gunicorn